LPS-56933 Rename MB module packages from `comment` to `comments`
