+++ modules/apps/message-boards/message-boards-comment/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Message Boards Comment
-Bundle-SymbolicName: com.liferay.message.boards.comment
+Bundle-Name: Liferay Message Boards Comments
+Bundle-SymbolicName: com.liferay.message.boards.comments
 Bundle-Version: 1.0.0
 Include-Resource: classes
