+++ modules/apps/message-boards/message-boards-comment/src/com/liferay/message/boards/comments/MBCommentManagerImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.comment;
+package com.liferay.message.boards.comments;
 
 import com.liferay.portal.kernel.comments.Comment;
 import com.liferay.portal.kernel.comments.CommentConstants;
