+++ modules/apps/message-boards/message-boards-comment/src/com/liferay/message/boards/comments/MBDiscussionImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.comment;
+package com.liferay.message.boards.comments;
 
 import com.liferay.portal.kernel.comments.Discussion;
 import com.liferay.portal.kernel.comments.DiscussionComment;
